Padrão Domain Events do DDD - Publicando e consumindo eventos
