+++ src/main/java/com/course/springfood/domain/event/PedidoConfirmadoEvent.java
@@ -1,0 +1,14 @@
+package com.course.springfood.domain.event;
+
+import com.course.springfood.domain.model.Pedido;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class PedidoConfirmadoEvent {
+
+    private Pedido pedido;
+
+}
