+++ src/main/java/com/course/springfood/domain/listener/NotificacaoClientePedidoCanceladoListener.java
@@ -1,0 +1,32 @@
+package com.course.springfood.domain.listener;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.event.TransactionalEventListener;
+
+import com.course.springfood.domain.event.PedidoCanceladoEvent;
+import com.course.springfood.domain.model.Pedido;
+import com.course.springfood.domain.service.EnvioEmailService;
+import com.course.springfood.domain.service.EnvioEmailService.Mensagem;
+
+@Component
+public class NotificacaoClientePedidoCanceladoListener {
+
+    @Autowired
+    private EnvioEmailService envioEmail;
+
+    @TransactionalEventListener
+    public void aoCancelarPedido(PedidoCanceladoEvent event) {
+        Pedido pedido = event.getPedido();
+
+        var mensagem = Mensagem.builder()
+                .assunto(pedido.getRestaurante().getNome() + " - Pedido cancelado")
+                .corpo("pedido-cancelado.html")
+                .variavel("pedido", pedido)
+                .destinatario(pedido.getCliente().getEmail())
+                .build();
+
+        envioEmail.enviar(mensagem);
+    }
+
+}
