+++ src/main/java/com/course/springfood/domain/listener/NotificacaoClientePedidoConfirmadoListener.java
@@ -1,0 +1,32 @@
+package com.course.springfood.domain.listener;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.event.TransactionalEventListener;
+
+import com.course.springfood.domain.event.PedidoConfirmadoEvent;
+import com.course.springfood.domain.model.Pedido;
+import com.course.springfood.domain.service.EnvioEmailService;
+import com.course.springfood.domain.service.EnvioEmailService.Mensagem;
+
+@Component
+public class NotificacaoClientePedidoConfirmadoListener {
+
+    @Autowired
+    private EnvioEmailService envioEmail;
+
+    @TransactionalEventListener
+    public void aoConfirmarPedido(PedidoConfirmadoEvent event) {
+        Pedido pedido = event.getPedido();
+
+        var mensagem = Mensagem.builder()
+                .assunto(pedido.getRestaurante().getNome() + " - Pedido confirmado")
+                .corpo("pedido-confirmado.html")
+                .variavel("pedido", pedido)
+                .destinatario(pedido.getCliente().getEmail())
+                .build();
+
+        envioEmail.enviar(mensagem);
+    }
+
+}
